Update to remote master

+++ scripts/to_yaml.py
@@ -1,58 +1,97 @@
 #!/usr/bin/env python
-
 import glob
 import json
 import os
-import re
-import sys
 import uuid
-import yaml
-import yamlordereddictloader
+import click
 from collections import defaultdict, OrderedDict
-from utils import reformat_phone_number, get_data_dir
-
-# set up defaultdict representation
-from yaml.representer import Representer
-yaml.add_representer(defaultdict, Representer.represent_dict)
+from utils import (reformat_phone_number, reformat_address, get_data_dir, get_jurisdiction_id,
+                   dump_obj)
 
 
-def all_people(dirname):
-    memberships = defaultdict(list)
-    organizations = {}
-
-    for filename in glob.glob(os.path.join(dirname, 'organization_*.json')):
-        with open(filename) as f:
-            organization = json.load(f)
-            organizations[organization['_id']] = organization
-
-    for filename in glob.glob(os.path.join(dirname, 'membership_*.json')):
-        with open(filename) as f:
-            membership = json.load(f)
-            membership['organization'] = organizations.get(membership['organization_id'])
-            memberships[membership['person_id']].append(membership)
-
-    for filename in glob.glob(os.path.join(dirname, 'person_*.json')):
-        with open(filename) as f:
-            person = json.load(f)
-            person['memberships'] = memberships[person['_id']]
-            yield person
-
-
-def filename_for_person(person):
-    id = person['id']
-    name = person['name']
-    name = re.sub('\s+', '-', name)
-    name = re.sub('[^a-zA-Z-]', '', name)
-    return f'{name}-{id}.yml'
-
-
-def postprocess_link(link):
+def process_link(link):
     if not link['note']:
         del link['note']
     return link
 
 
-def postprocess_person(person, jurisdiction_id):
+def ocd_uuid(type):
+    return 'ocd-{}/{}'.format(type, uuid.uuid4())
+
+
+def process_dir(input_dir, output_dir, jurisdiction_id):
+    person_memberships = defaultdict(list)
+    # map both names & ids to people objects
+    people_lookup = {}
+    committees_by_id = {}
+
+    # build list of committees
+    for filename in glob.glob(os.path.join(input_dir, 'organization_*.json')):
+        with open(filename) as f:
+            org = json.load(f)
+
+        if org['classification'] == 'committee':
+            committees_by_id[org['_id']] = process_org(org, jurisdiction_id)
+
+    # collect memberships
+    for filename in glob.glob(os.path.join(input_dir, 'membership_*.json')):
+        with open(filename) as f:
+            membership = json.load(f)
+
+        if membership['organization_id'] in committees_by_id:
+            committees_by_id[membership['organization_id']]['memberships'].append(membership)
+        else:
+            if membership['person_id'].startswith('~'):
+                raise ValueError(membership)
+            person_memberships[membership['person_id']].append(membership)
+
+    # process people & store people by ID for committees
+    for filename in glob.glob(os.path.join(input_dir, 'person_*.json')):
+        with open(filename) as f:
+            person = json.load(f)
+
+        scrape_id = person['_id']
+        person['memberships'] = person_memberships[scrape_id]
+        person = process_person(person, jurisdiction_id)
+        people_lookup[scrape_id] = person
+        people_lookup[person['name']] = person
+
+        dump_obj(person, output_dir=os.path.join(output_dir, 'people'))
+
+    # resolve committee parents and members and write them out
+    for org in committees_by_id.values():
+        if org['parent'].startswith('~'):
+            org['parent'] = json.loads(org['parent'][1:])['classification']
+
+        org['memberships'] = [process_committee_membership(m, people_lookup)
+                              for m in org['memberships']]
+
+        dump_obj(org, output_dir=os.path.join(output_dir, 'organizations'))
+
+
+def process_committee_membership(membership, people_lookup):
+    result = OrderedDict()
+    if membership['person_id'].startswith('~'):
+        try:
+            result['id'] = people_lookup[membership['person_name']]['id']
+        except KeyError:
+            # there are many unresolved people for all sorts of reasons,
+            # we'll see them in the lint
+            pass
+    else:
+        result['id'] = people_lookup[membership['person_id']]['id']
+
+    result['name'] = membership['person_name']
+    if membership['role'] != 'member':
+        result['role'] = membership['role']
+    if membership['start_date']:
+        result['start_date'] = membership['start_date']
+    if membership['end_date']:
+        result['end_date'] = membership['end_date']
+    return result
+
+
+def process_person(person, jurisdiction_id):
     optional_keys = (
         'image',
         'gender',
@@ -68,15 +107,14 @@ def postprocess_person(person, jurisdiction_id):
     )
 
     result = OrderedDict(
-        id=str(uuid.uuid4()),        # let's use uuid4 for these, override pupa's
+        id=ocd_uuid('person'),
         name=person['name'],
         party=[],
         roles=[],
-        links=[postprocess_link(link) for link in person['links']],
+        links=[process_link(link) for link in person['links']],
         contact_details=[],
         # maybe post-process these?
-        sources=[postprocess_link(link) for link in person['sources']],
-        committees=[],
+        sources=[process_link(link) for link in person['sources']],
     )
 
     contact_details = defaultdict(lambda: defaultdict(list))
@@ -84,6 +122,8 @@ def postprocess_person(person, jurisdiction_id):
         value = detail['value']
         if detail['type'] in ('voice', 'fax'):
             value = reformat_phone_number(value)
+        elif detail['type'] == 'address':
+            value = reformat_address(value)
         contact_details[detail['note']][detail['type']] = value
 
     result['contact_details'] = [{'note': key, **val} for key, val in contact_details.items()]
@@ -91,35 +131,31 @@ def postprocess_person(person, jurisdiction_id):
     # memberships!
     for membership in person['memberships']:
         organization_id = membership['organization_id']
-        if organization_id.startswith('~'):
-            org = json.loads(organization_id[1:])
-            if org['classification'] in ('upper', 'lower'):
-                post = json.loads(membership['post_id'][1:])['label']
-                result['roles'] = [
-                    {'chamber': org['classification'], 'district': post,
-                     'jurisdiction': jurisdiction_id}
-                ]
-            elif org['classification'] == 'party':
-                result['party'] = [
-                    {'name': org['name']}
-                ]
-        elif membership['organization']:
-            result['committees'].append({
-                'name': membership['organization']['name'],
-            })
-        else:
+        if not organization_id.startswith('~'):
             raise ValueError(organization_id)
+        org = json.loads(organization_id[1:])
+        if org['classification'] in ('upper', 'lower'):
+            post = json.loads(membership['post_id'][1:])['label']
+            result['roles'] = [
+                {'type': org['classification'], 'district': post,
+                 'jurisdiction': jurisdiction_id}
+            ]
+        elif org['classification'] == 'party':
+            result['party'] = [
+                {'name': org['name']}
+            ]
 
     for key in optional_keys:
         if person.get(key):
             result[key] = person[key]
 
     # promote some extras where appropriate
+    extras = person.get('extras', {}).copy()
     for key in person.get('extras', {}).keys():
         if key in optional_keys:
-            result[key] = person['extras'].pop(key)
-    if person.get('extras'):
-        result['extras'] = person['extras']
+            result[key] = extras.pop(key)
+    if extras:
+        result['extras'] = extras
 
     if person.get('identifiers'):
         result['other_identifiers'] = person['identifiers']
@@ -127,18 +163,24 @@ def postprocess_person(person, jurisdiction_id):
     return result
 
 
-def process_people(input_dir, output_dir, jurisdiction_id):
-    for person in all_people(input_dir):
+def process_org(org, jurisdiction_id):
+    return OrderedDict(
+        id=ocd_uuid('organization'),
+        name=org['name'],
+        jurisdiction=jurisdiction_id,
+        parent=org['parent_id'],
+        classification=org['classification'],
+        links=[process_link(link) for link in org['links']],
+        sources=[process_link(link) for link in org['sources']],
+        memberships=[],
+    )
 
-        person = postprocess_person(person, jurisdiction_id)
-        filename = filename_for_person(person)
 
-        with open(os.path.join(output_dir, filename), 'w') as f:
-            yaml.dump(person, f, default_flow_style=False, Dumper=yamlordereddictloader.Dumper)
-
-
-if __name__ == '__main__':
-    input_dir = sys.argv[1]
+@click.command()
+@click.argument('input_dir')
+@click.option('--reset/--no-reset', default=False)
+def to_yaml(input_dir, reset):
+    # TODO: remove reset option once we're in prod
 
     # abbr is last piece of directory name
     abbr = None
@@ -148,17 +190,17 @@ if __name__ == '__main__':
             break
 
     output_dir = get_data_dir(abbr)
+    jurisdiction_id = get_jurisdiction_id(abbr)
 
-    if abbr == 'dc':
-        jurisdiction_id = 'ocd-jurisdiction/country:us/district:dc'
-    elif abbr in ('vi', 'pr'):
-        jurisdiction_id = f'ocd-jurisdiction/country:us/territory:{abbr}'
-    else:
-        jurisdiction_id = f'ocd-jurisdiction/country:us/state:{abbr}'
+    for dir in ('people', 'organizations'):
+        try:
+            os.makedirs(os.path.join(output_dir, dir))
+        except FileExistsError:
+            if reset:
+                for file in glob.glob(os.path.join(output_dir, dir, '*.yml')):
+                    os.remove(file)
+    process_dir(input_dir, output_dir, jurisdiction_id)
 
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        for file in glob.glob(os.path.join(output_dir, '*.yml')):
-            os.remove(file)
-    process_people(input_dir, output_dir, jurisdiction_id)
+
+if __name__ == '__main__':
+    to_yaml()
